Change ID allocator to round robin strategy
Tunnel IDs allocated to pods have a more dynamic lifecycle than IDs
allocated to networks. Lets change the allocation strategy to round
robin to reduce the posibility of potential collisions when a
de-allocated tunnel ID is allocated to a pod while still being used
by one of the zones that might have not yet process the original pod
termination event.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>