Fixes race in delete legacy hybrid route policies
This test adds 2 legacy hybrid routes in the old format and then 1 new
one and runs the stale cleanup code to verify that legacy ones get
removed. By starting APB controller it will also try to run repair and
potentially remove the new one becuase it does not match a real pod in
the cluster. There is no real value in having APB in this test.

Signed-off-by: Tim Rozet <trozet@redhat.com>