Fixes ovnkube-controller not waiting for zone
Ovnkube-controller would exit if it did not see the zone configured in
NBDB on the first try because the wait loop was returning error when it
did not find the zone, rather than continuing to poll. This fixes it and
adds some more clear logging on how long it is taking for this wait to
finish.

Signed-off-by: Tim Rozet <trozet@redhat.com>