ovn-ic: node update missing static routes
When node from a remote zone is updated, we only perform the
actual update when necessary. This commit improved the logic
for doing the remote update in cases where the subnets of the
remote node change. That is particularly needed when node
changes from ipv4 to dual stack (ipv4 + ipv6)

Reported-at: https://issues.redhat.com/browse/SDN-3993
Signed-off-by: Flavio Fernandes <flaviof@redhat.com>