Adds suppressed type errors to ignore certain errors
With IC, we now get a lot of "errors" spammed in the log about remote
pods not finding IP addresses because we have to wait for update events
by other nodes to assign their IPs. The error typically looks like:

error: failed to obtain IPs to add remote pod

and is flooding our logs:
https://gcsweb-ci.apps.ci.l2s4.p1.openshiftapps.com/gcs/origin-ci-test/logs/release-openshift-origin-installer-launch-aws-modern/1673971066382323712/artifacts/launch/gather-extra/artifacts/pods/openshift-ovn-kubernetes_ovnkube-node-6j65p_ovnkube-controller.log

Since we know during the pod add event, and maybe some subsequent
updates we will not have the pod IP and it is expected for a remote pod,
we simply do not log this as an error, but still treat it as one. When
the retry framework reaches its max failed attempts to get the pod IP
and fails, we then log the final error there.

Note the retry framework will still log failed intermediate attempts
as Info level, since they are going to retry.

Signed-off-by: Tim Rozet <trozet@redhat.com>