Fix egressip test not waiting for update processing before delete
This egressip test is updating and then deleting an egressip. The
expectations in between do not correctly verify if the update had been
processed. Sometimes the update would not be processed at all because we
get the delete event before that.

Then when we process the delete, we only look at the deleted egress ip
status items but not at any other status items the pod would have been
previously assigned. In this specific case those other status items are
not cleared and the standby egressip does not take over.

This fixes the UT to correctly verify that the update has been process.
But there might be another bug to fix in the egressip logic.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>