Ensure ovnkube-controller does not go remote->local
At ovnkube-controller start up, it may attempt to start adding resources
for a node that does not yet have its annotation, but should be
considered local. This is especially the case in OpenShift when going
from phase1 -> phase 2 upgrade and ovnkube-node on the node will
annotate its zone id. In this case ovnkube-controller will start too
early and program NBDB resources as remote for the node. Then when the
node is annotated, it will program NBDB resources as local.

Since we cannot be sure that all remote feature configuration in NBDB is
removed when going from remote -> local we need to avoid this case. It
also makes no sense to program NBDB when there are no nodes in the zone.

This commit adds a wait to make sure at least one node in the cluster
is in the ovnkube-controller's managed zone.

Thanks @numansiddique for the idea.

Signed-off-by: Tim Rozet <trozet@redhat.com>