Fixes race across node workers to create transit switch
Previously the behavior was for every remote add across 15 threads to
try to create or update the transit switch. The transit switch is a
singleton which does not need to be updated on every remote node add.
Consequently by doing this it could cause high delays in nodes getting
added for IC. This is because one or more threads could read the
libovsdb cache at the same time and determine the switch does not exist,
then make insert operations. These insert operations are guarded by a
libovsdb wait, and the timeout on retries per transaction is 10 seconds.
Compounded into this delay is the retry framework timer which is 30
seconds. This could result in nodes delaying to be added by up to 40
seconds.

This patch changes the behavior so that the transit switch is configured
during ovnkube-controller init phase. In order to do this, we may need
to wait for the cluster manager to assign the network ID to the node and
annotate it.

Signed-off-by: Tim Rozet <trozet@redhat.com>