e2e multi-homing: fix uneeded replacement of network name
PR#3609 removed the suffix `_br-localnet` from the localnet network
name used in the bridge mappings, but also inadvertedly removed the
replacement of `-` and `/` characters form the controller code but not
from the tests.

Let's remove them from tests as well as they don't look troublesome
neither as network name in the logical switch port options nor in the
OVS bridge mappings.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>