Replace child stop channels with cancelable contexts
Commit 5d6b136 added child stop channels to stop the network policy
handlers independently from the network controller when the policy is
deleted while also stopping them if the network controller is stopped.

Unfortunately when both things happen at the same time, one of those
events will end up attempting to close a closed channel which will panic.

Introduce a CancelableContext utility that will wrap a cancelable context
that can be chained to achieve the same effect.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>