kubevirtl: Replace node only at migration tests
After live migrating a VM the original node can be shutdown and a new
one can appear taking over the node subnet. There were some unit tets
for that but they exercise this part for non live migration scenarios
creating a race condition between cleanup and addLogicalPort.

This change run that part of the tests only for live-migration
scenarios.

Signed-off-by: Enrique Llorente <ellorent@redhat.com>