Make OVS pinning period shorter
The overall system architecture means that it would take more than 15 seconds to propagate the affinity to OVS. That is too long and the workloads are not expecting that.

Kubelet's CPU manager itself is configured for 5 seconds in the deployments this is supposed to be used with and so adding 10 more is just too much.

Reducing that to one second should still cause no extra burden on the system and at the same time make it more predictable for the users.

Signed-off-by: Martin Sivak <msivak@redhat.com>