Move ExtractPortAddresses&GetLRPAddrs into libovsdb/utils
This commit creates a new pkg libovsdb/utils and moves
some of the helper util functions into that package
instead of keeping them in main utils. This is because
importing libovsdbops from util package is not a neat
way since in future if libovsdbops package wants to use
util functions it will lead to circular dependency.

SO let's create libovsdb/util package and move wrapper
functions that use libovsdbops package from utils into this.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>