Fixes service controller + node tracker start up race
When service controller starts up along side node tracker, it is
possible that it attempts to sync a service without having all of the
nodes in the cluster. In new installations this is fine, but in a steady
state cluster where a new ovnkube-controller takes over, or is restarted
it can cause service outage.

This happens because there may already be a load balancer configured for
a specific node in OVN, then when the ovnk pod is restarted, it may sync
the service before it has rebuilt its node cache. The consequence is the
LB will be removed from OVN for a period of time, causing traffic
disruption. This was noticed in a downstream OCP CI run.

This commit fixes the problem by ensuring that the node tracker has
synced its initial objects first, before allowing services and endpoint
handlers to be added and their respective workers to be started.

Signed-off-by: Tim Rozet <trozet@redhat.com>