Fix encap port deadlock
If geneve encap port is different from the default,
then we configure the SBDB chassis with the new port.
After this we proceed to annotate the zone-name in
ovnkube-node.
However in IC deployments, ovnkube-controller does not
come up until the zone-name annotation is tagged on the
nodes: https://github.com/ovn-org/ovn-kubernetes/blob/4bfacf905e84e9e1e37e3a9dda5b5321df22bc4f/go-controller/pkg/network-controller-manager/network_controller_manager.go#L337
This means ovnkube-controller is waiting on ovnkube-node
to annotate stuff before starting to configure NBDB stuff
and SBDB chassis creation won't happen because of that.

On the other hand, ovnkube-node is timing out trying to
find the sbdb chassis to set the encap port since chassis
is not existing in sbdb yet.

We end up in a deadlock situation.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>