Delete legacy DGP mac binding using client native API
Delete using the native client API instead of model client API to avoid
having to monitor & cache the SBDB mac binding table. As opposed to the
model client API, which would result in noop if there is no cache hit,
the native client API will successfully build a delete OP regardless of
whether there is a cache hit or not. The operation is idempotent and
there will be no error if there is nothing to delete.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>