EIP TC: stop flake by ensuring EIP spec and status are equal
The EIP specified in the EIP spec must equal what is set
in the EIPs status item otherwise, CM EIP controller,
will delete the EIP and attempt to reassign it.
The purpose of this test was to ensure this does not
occur so make the status EIP the same lexicographically
to the EIP specified in the spec.

Signed-off-by: Martin Kennelly <mkennell@redhat.com>