clearInitialNodeNetworkUnavailableCondition() called on node update
when this used to be called from a master pod
clearInitialNodeNetworkUnavailableCondtion() was called during both add
and update. It seems as though there are some instances where there is a
delay between when the node is created and when gcp adds the taint,
clearing on unpdate will ensure that the taint is removed.

Signed-off-by: Jacob Tanenbaum <jtanenba@redhat.com>