Allocate IPs of completed pods on startup
There is a race condition on startup when a scheduled pod is allocated
an IP that is also annotated on a completed pod by which the delete flow
of the completed pod will check if any other pods are using that IP just
before the scheduled pod is annotated, missing that and thus
de-allocating the IP, leading to the possibility of a third pod being
allocated that same IP and ending up with duplicate IPs for two different
pods.

By allocating the IPs of completed pods on startup, we prevent other
scheduled pods from being allocated the same IPs.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>