Don't release completed pods sharing IPs with running pods
On startup, multiple pods might be annotated with the same IP as long as
just one of them is running. Thus, we cannot release the IPs for the
completed ones as that would de-allocate the IP which the running pod is
using.

To avoid that, track which pods are annotated with the same IPs on
startup and make sure that those IPs are only released with the
appropriate pods, that is, with a running pod or with the last completed
pod if there is no running pod sharing that IP. Do this tracking per
NAD to cover the corner case where a pod might be released for one NAD
but others in error cases.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>