Track released live migrated pods on unassigned subnets
A live migrated pod might have an IP annotated from a subnet assigned
to a different node than the one is running on or even to a node that has
since been deleted and not re-assigned to another node. We also need to
track the IPs of these pods in case they are shared among completed pods
or running pods to make sure they are released for the correct pod.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>