change the name of host-addresses annotation to host-cidrs
commit: 8847618bbf2feeb42c30375a6de6f91afe011d91 changed the
host-addresses annotation to include a subnet mask. This causes updates
to experiance a network disruption as masters and nodes expect different
data to be in the host-addresses annotation.

Changing the name of the host-addresses annotation to host-cidrs will
eliminate the network disruption because nodes updating will not remove
the host-addresses annotation and once upgraded masters will do the
right thing with the host-cidrs annotation

Co-authored-by: Tim Rozet <trozet@redhat.com>
Signed-off-by: Jacob Tanenbaum <jtanenba@redhat.com>