Implement port security for OVN patch port and GR
There can be a case where the MAC address might change of the physical
interface attached to OVS. One common way this can happen is with a
bond. The MAC address of the bond may change depending on which slave
comes up first.

When this scenario happens, the host will now be using a new MAC, while
the OVN GR has the old MAC. This can cause hosts outside of the cluster
to constantly update their IP neighbor table with the wrong MAC for the
node, and cause traffic outage.

In order to prevent such a scenario, ensure that OVNK does not allow
traffic sent by OVN GR that does not match the current MAC address.

Signed-off-by: Tim Rozet <trozet@redhat.com>