Ensure required flows are present at node start
In the scenario where an OVN GR may have the wrong mac, and traffic
disruption is present, the OVNK node process may stall at trying to
start the watch factory and wait for informer caches to sync. At this
point in time, ovnkube node is not able to start its OpenFlow manager
and program the needed flows to block the OVN GR from poisoning external
ip neighbor entries.

This commit adds a boostrap function which attempts to install basic
flows before anything else is done on the node during boot time. This is
a temporary flow installation that is only done when there is just a
single NORMAL flow in the external bridge and is overriden once OF
Manager starts.

Note this intentionally does not write the bootstrap flows if OVNK has
been killed, but the OVS flows from the previous run remain. It is only
for cases where OVS has no previously programmed flows, like on node
boot up.

Signed-off-by: Tim Rozet <trozet@redhat.com>