kubevirt: Ignore completed virt-launcher pods
When a live migration fails at completed virt-launcher pod with newer
creation timestamp is lingering in the system so the virt-launcher that
is really running the VM is hidden by it and routes are wrongly
reconstructed. This change take into account if the pod is completed to
ignore it marking it as stale.

Signed-off-by: Enrique Llorente <ellorent@redhat.com>