Account for the new VIP address label
The previous ${interface}:vip label on keepalived addresses yielded an error whenever the total label length exceeded the allowed limit of 15 characters. As a result, we simplified the label to be just "vip", which is sufficient for our purposes: https://github.com/openshift/machine-config-operator/pull/4040
Let's keep the call to "HasSuffix" to be compatible with both implementations.

Signed-off-by: Riccardo Ravaioli <rravaiol@redhat.com>