Add interface filter for drop-forwarding feature
The drop-forwarding feature by default creates drop rules for all
physical and VLAN interfaces. When --drop-forwarding-interfaces is
provided a valid regex, use that filter expression instead to determine
physical interfaces. This is mainly needed for E2E testing (inside
docker containers, the `eth.*` interfaces are all of type veth, so we
need a switch to overrided the selection mechanism) but might also be
useful in production when administrators want more control.

Signed-off-by: Andreas Karis <ak.karis@gmail.com>