Deprecate IPv6 host masq IP in-order to use it for svc only
Currently, the host masq IP is selected as src IP when a workload
residing in the host net ns attempts to connect to
a svc IP.

But users may not config any routes that match a particular
destination. This may happen if a user configures addresses
on a link with noprefixroute or if the destination is not
covered by any route.

For IPv6, there is an increased chance of the host masq IP
being selected as src IP for destinations that do not fall
within the svc CIDR if a user does not have appropriate
routes configured to ensure the host masq IP is not selected
and we have to use the following algorithm for V6 addrs:
(rfc-3484)

1. Prefer same address.
2. Prefer appropriate scope.
3. Avoid deprecated addresses.
4. Prefer home addresses.
5. Prefer outgoing interface
6. Prefer matching label.
7. Prefer public addresses.
8. Use longest matching prefix.

If theres a tie break, linux uses the last address added,
and that can end up being the host masq IP.

We want to only use the host masq IP if the dst is a
a service IP and never use it otherwise.

We cannot change the scope of the address because the host
IP masq IP falls within the global scope range as defined
by IANA special registry and it wouldnt totally solve
the problem anyway, just make it less likely to occur.

Deprecating the address was not meant for this purpose
but it partially achieve our goal of only using this IP
as src IP for svcs as long as we have the correct service
routes configured with the correct src IP.

I state partially, because step 3. in the algo, only
states to "avoid" dep addresses.

Issue: https://issues.redhat.com/browse/OCPBUGS-17207

Signed-off-by: Martin Kennelly <mkennell@redhat.com>