Fix informer_test: adds existing pod and processes an update event
When starting the informer factory and the worker thread, a delay
may occur between the informer being deemed synced and the worker
completely starting. Due to this, the test could fail as add and
update pod events might coalesce into a single event for the
worker. This change proposes a fix that ensures the worker is fully
initialized before updating the pod.

Step used to reproduce this issue. Should fail after a few iterations.

$ cd go-controller ; \
  while : ; do \
  go test -mod=vendor -test.v \
  -race github.com/ovn-org/ovn-kubernetes/go-controller/pkg/informer \
  -ginkgo.v \
  -ginkgo.focus='adds existing pod and processes an update event' \
  -ginkgo.reportFile /dev/null && echo ok || break ; \
  sleep 1 ; done

Fixes: https://github.com/ovn-org/ovn-kubernetes/issues/3946
Reported-at: https://issues.redhat.com/browse/SDN-4181
Signed-off-by: Flavio Fernandes <flaviof@redhat.com>