Use full datapath name when querying datapaths with ovs-appctl
When OVN-Kubernetes lists all datapaths with 'ovs-appctl dpctl/dump-dps',
it stores both a datapath's name and type. But it then continues to use
only each datapath's name when querying 'ovs-appctl dpctl/show'.

For system datapaths this works because 'ovs-appctl dpctl/show' will
assume the 'system' type if no type has been set. However, this breaks
when using other datapath types such as 'netdev', causing 'ovs-appctl'
to fail with:

  ovs-vswitchd: opening datapath ovs-netdev failed (No such device)

This change keeps using a datapath's name as an identifier for metrics
but includes the type when querying ovs-appctl.

Whether a datapath's name without a type is sufficient for identifying
datapaths uniquely in OpenShift and Open vSwitch is left for future
work.

Signed-off-by: Jakob Meng <code@jakobmeng.de>