EIP multi NIC: in LGW mode, skip IP rules when dst is node IP
In LGW mode, we need to skip IP rules when dst is a node IP.
Within ovn_cluster_router LRPs, we mark pkts that are destined for
a node IP.

However, when we use the pkt mark in an IP rule, it fails reverse path
filtering for the response path, therefore we only save the mark if
its the specific mark for node IP. We restore the mark for the response path.

We also must tell the kernel to use the pkt mark when performing reverse
path filtering. [1]

[1] https://docs.kernel.org/networking/ip-sysctl.html (search for
src_valid_mark)

Signed-off-by: Martin Kennelly <mkennell@redhat.com>