EIP multi NIC: copy routes from main routing table
Previous to this change, we added just an any or default
route to the routing table.
With this change, we copy any routes associated with the egress
link to the new routing table and if there is no any or default
route, we add it.

Also, routes within in a custom routing table are mapped directly
with a link, however multiple egress IPs can be associated with a
link. Do not delete routes in the custom routing table if an
existing egress IP is assigned to the link.

Also, within the repair func, previous to this change,
we looked to see if egress IPs we expect are present on the
system, but we never checked if the egress IP is assigned to
the correct link. With this change, we ensure that the egress IP
is assigned to the correct link and if it not, remove it.

Signed-off-by: Martin Kennelly <mkennell@redhat.com>