github: Replace deprecated set-output.
'set-output' was deprecated [1] long ago and supposed to be completely
removed last summer, but GitHub is keeping it for now due to
widespread use.  It throws a warning on every build and GitHub may
decide to actually stop supporting these at any time.

Update the workflow to use a modern GITHUB_OUTPUT environment file
instead.

[1] https://github.blog/changelog/2022-10-11-github-actions-deprecating-save-state-and-set-output-commands/

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>