github: Cache go dependencies to avoid warnings.
New setup-go@v5 attempts to cache dependencies by default.  However,
the default path it uses is go.sum in the root directory.  This
triggers a warning, since it doesn't exist:

  Restore cache failed: Dependencies file is not found in
  /home/runner/work/ovn-kubernetes/ovn-kubernetes.
  Supported file pattern: go.sum

Specify a path to all .sum files we have in the repository to make
the setup-go happy.  This should in theory make the builds a touch
faster.

Repo checkout moved before setup-go, so the paths are available.

Also removed the strange last_run_status logic.  There is no such
step anywhere.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>