Periodically check the ovnkube-node certificate is not expired
Once the initial certificate is retrieved, run a periodic check to ensure it is not expired.
The certificate can become invalid when the time changed unexpectedly(e.g. cluster was suspended).

Additionally, when waiting for the initial certificate use the one from cert manager to avoid
a scenario in which an expired certificate is accepted as valid.

Signed-off-by: Patryk Diak <pdiak@redhat.com>