Fix HO unit test flake
The node ID 1 and in correspondence, the first ip of the join subnet, is
reserved for the OVN cluster router. This HO unit test was trying to
use that ID for a node and the test would flake depending on whether
cluster manager was able to fix the situation before the test ran or
not.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>