kubevirt, e2e: Use kube api service to test n/s
Using depending on external domain to check n/s traffic and name
resolution introduce some flakiness since the domain can be down or
reject CI.

This change replace the external domain with the kube api service fqdn
so everything is setup by the cluster.

Also to make this work the guest VM is configured to use NetworkManager
for name resolution instead of systemd resolved.

Signed-off-by: Enrique Llorente <ellorent@redhat.com>