Adds MEG migration e2e test
New test case where egress traffic is sent to a static destination and
source port, which results in a single gateway chosen. Afterwards, the
MAC is modified of the gateway, and an GARP/unsolicted neighbor
advertisement is sent. Then more egress traffic is sent and verified
that it still goes to the same gateway and is not dropped.

Signed-off-by: Tim Rozet <trozet@redhat.com>