Add dynamic BFD rate limiter
This commit adds logic to dynamically
determine the meterBand for BFD packets.
By default it is set to 50 and every time
we add a new BFD session this will get
adjusted.

NOTE: We have not added logic to reduce the
band as that doesn't seem to make much sense.
Goal is to catch the upper bound, not lower bound.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>