Remove the configmap based OVN topology version reporting/detection
The current implementation where we report the topology from every
node is not going to work in interconnect.
Considering that the topology version has not change for a while
remove the code as we do not forsee upgrades from such old versions.
We can revisit a similar approach in the future if there is a need.

Signed-off-by: Patryk Diak <pdiak@redhat.com>