kubevirt: Disable guest console log container
There is a bug [1] at kubevirt where source and target pods are not ready
after successful live migration, this make endpoints slices move not
ready and ovn-k do some ovs/ovn cleanup that breaks connections.

This change reduce the change to have a not ready source pod by
disabling one of the optional containers that is more eager to fail.

[1] https://issues.redhat.com/browse/CNV-38604

Signed-off-by: Enrique Llorente <ellorent@redhat.com>